Refactor namespaces for better code organization
Reorganized namespaces for test and analyzer classes to improve
code structure and maintainability. Specific changes include:
- Moved analyzer test classes to Analyzers.Analyzers namespace.
- Moved code fix test classes to Analyzers.CodeFixes namespace.
- Moved source generator test classes to Analyzers.SourceGenerators namespace.
- Made EntityConstructorAnalyzer class sealed.
